:white_check_mark: refactor tests
