+++ test/integration/resolve-base-urls/fixtures.ts
@@ -1,0 +1,9 @@
+import { GlobalFixtures } from '../fixtures';
+
+export class Fixtures extends GlobalFixtures {
+  public static readonly defaultMatchValue = '*://*/*';
+
+  public static readonly httpsMatchValue = 'https://*/*';
+
+  public static readonly httpsIncludeValue = 'https://*';
+}
